Store refactoring
