+++ public/index.html
@@ -6,11 +6,11 @@
     <link rel="shortcut icon" type="image/x-icon" href="/favicon.ico" />
     <!-- Compiled and minified CSS -->
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/semantic-ui/2.4.1/semantic.min.css" />
-    <title>React Todo App</title>
+    <title>Project Board Management</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-    <script src="../src/index.tsx" defer></script>
+    <script src="../src/index.js" defer></script>
   </body>
 </html>
